new function

+++ Course_with_angular/Models/Rate.cs
@@ -10,6 +10,8 @@ namespace Course_with_angular.Models
         public int Id { get; set; }
         public int ProjectId { get; set; }
         public string UserId { get; set; }
-        public float Mark { get; set; }
+        public int Mark { get; set; }
+
+
     }
 }
